stop trying to send e-mails to admins
PAY ATTENTION:

Since we do not have properly set reverse DNS record, e-mails are not
accepted by Gmail smtp server and are deferred. That means they are
returned to postfix queue and our production server tries to send them
later. As this queue is growing (>> 10,000 e-mails), it sends a really
large number of e-mails per minute and it causes exceeding of Gmail
limits and Gmail stops receiving all e-mails.

Before turning sending of e-mails on, we have to fix the issue with DNS
record.